Air Pollution Monitoring System Based On IOT
Air is essential for living organisms in their daily lives, as they use it as a tool for breathing. In terms of health, the air needed by living organisms is clean air. However, with the advancement of time, the supply of clean air is becoming increasingly scarce due to contamination by human activities, resulting in this clean air becoming contaminated by toxic chemicals. This can have negative impacts on human health. Therefore, a system can be created to monitor clean air in an area or room if there are toxic chemicals present. This system will be realized through the Internet of Things (IoT) in its operation. Sensors to be used for this system include MQ2, MQ4, MQ7, and MQ135 sensors as gas detectors, detecting gas, methane, carbon monoxide, and carbon dioxide in an environment or room to monitor pollution present in an area. If toxic chemicals reach the specified limit, the buzzer will sound, and the servo motor will automatically rotate as a fan to remove the toxic chemicals, while another servo motor will function to pull oxygen from the outside into the room. This system uses the ESP32 as the microcontroller connected to Node-RED