System Automatic Sliding Door Security System
Generally, security systems for office doors typically use conventional keys. The more office rooms there are, the more keys are needed. The challenges faced include difficulty in finding keys and losing them. As time progresses, one way to overcome these problems is by creating an automatic door locking system using fingerprints. This system aims to enhance security and simplify door access. It is constructed using an Arduino Nano as the microcontroller, equipped with a fingerprint sensor for fingerprints, an LCD serving as output displaying text, a buzzer functioning as an output emitting sound, and a solenoid door lock for locking.